third_party/WebKit/LayoutTests/external/wpt/payment-request/allowpaymentrequest/removing-allowpaymentrequest.https.sub-expected.txt

This is a testharness.js-based test.
FAIL PaymentRequest removing allowpaymentrequest after load and then navigatingassert_equals: after navigation expected "Exception" but got "Success"
Harness: the test ran to completion.
